Delete all activities and replace them with fully composable files
