+++ app/src/main/java/com/jeeps/gamecollector/deprecated/AddGameActivityOld.java
@@ -86,7 +86,6 @@ public class AddGameActivityOld extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_add_game);
         context = this;
 
         //Change title
