+++ app/src/main/java/com/jeeps/gamecollector/deprecated/MainLibraryActivity.java
@@ -70,7 +70,6 @@ public class MainLibraryActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main_library);
 //        Toolbar toolbar = findViewById(R.id.toolbar);
 //        setSupportActionBar(toolbar);
