+++ app/src/main/java/com/jeeps/gamecollector/remaster/ui/MainActivity.kt
@@ -1,0 +1,25 @@
+package com.jeeps.gamecollector.remaster.ui
+
+import android.os.Bundle
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import com.jeeps.gamecollector.remaster.navigation.Main
+import com.jeeps.gamecollector.remaster.ui.base.BaseActivity
+import com.jeeps.gamecollector.remaster.ui.theme.AppTheme
+import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+
+@ExperimentalCoroutinesApi
+@AndroidEntryPoint
+class MainActivity : BaseActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            AppTheme {
+                Main()
+            }
+        }
+    }
+}
