--- app/src/main/res/layout/activity_add_game.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".remaster.ui.games.edit.AddGameActivity">
-
-    <androidx.compose.ui.platform.ComposeView
-        android:id="@+id/screen_compose"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:composableName="com.jeeps.gamecollector.remaster.ui.games.edit.AddGameActivityKt.AddGameScreenPreview"
-        />
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
