--- app/src/main/res/layout/activity_main_library.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.jeeps.gamecollector.deprecated.MainLibraryActivity">
-
-    <androidx.compose.ui.platform.ComposeView
-        android:id="@+id/screen_compose"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:composableName="com.jeeps.gamecollector.remaster.ui.gamePlatforms.GamePlatformsActivityKt.GamePlatformsScreenPreview"
-        />
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
