--- app/src/main/res/layout/activity_platform_library.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".remaster.ui.games.platformLibrary.GamesFromPlatformActivity">
-
-    <androidx.compose.ui.platform.ComposeView
-        android:id="@+id/screen_compose"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:composableName="com.jeeps.gamecollector.remaster.ui.games.platformLibrary.GamesFromPlatformActivityKt.GamesFromPlatformScreenPreview"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"/>
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
