--- app/src/main/res/layout/activity_stats.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.jeeps.gamecollector.deprecated.StatsActivity">
-
-    <androidx.compose.ui.platform.ComposeView
-        android:id="@+id/screen_compose"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:composableName="com.jeeps.gamecollector.remaster.ui.userStats.UserStatsActivityKt.UserStatsScreenPreview"
-        />
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
